Plaid API v2020-09-14 drop account creation
Plaid no longer supports the API-driven account setup process, instead
favoring their Plaid Link product, which is a HTML/JavaScript UI that
they control fully. This could be supported here, but it will be an
entirely new process involing running a webserver. It might be easier
to just follow the instructions in the erbridges/plaid2qif project
instead:
https://github.com/ebridges/plaid2qif#steps-to-authenticate-with-your-financial-institution